refactor(drift): single-source the FP/HJB drift coefficient via pde_coefficients.fp_drift_coefficient (#1420, G-017)
Promotes the G-017 drift-coefficient helper from a private fp_fdm_time_stepping function to a shared
public utility (pde_coefficients.fp_drift_coefficient) and applies it to the remaining independent
forks so the coefficient is single-sourced from the Hamiltonian's control_cost, not the independent
coupling_coefficient field:

- pde_coefficients.fp_drift_coefficient: 1/control_cost for a quadratic-MINIMIZE SeparableHamiltonian;
  falls back to coupling_coefficient for QuadraticMFGHamiltonian / non-Hamiltonian solves.
- fp_fdm_time_stepping: imports the shared util (behavior-identical to #1441).
- fp_fdm.solve_fp_step_adjoint_mode (FP side) and hjb_fdm.build_advection_matrix (HJB side) — the
  strict-adjoint pair — now both source via the shared util, so they stay consistent (A_fp = A_hjb.T
  preserved) and the SeparableHamiltonian case is corrected; QuadraticMFGHamiltonian is unchanged
  (falls back), so test_true_adjoint is unaffected.
- CHANGELOG [Unreleased] entry for the G-017 single-source-drift work.

Validation: CI gate (-m "not slow") 1541 passed, 0 failed; test_true_adjoint + G-017 pin + goldens
pass; the @slow MMS-EOC coupled test passes in 219s (byte-identical for it — coupling=1/lambda=1/control_cost).

Refs #1420, #1430. Follow-up to #1441. Remaining forks (fp_fvm, fp_particle) + synthetic-U deletion
+ SL 1/lambda (S0-03) are subsequent increments.
